+++ server/v2/commands.go
@@ -184,6 +184,9 @@ func topLevelCmd(ctx context.Context, use, short string) *cobra.Command {
 var appBuildingCommands = [][]string{
 	{"start"},
 	{"genesis", "export"},
+	{"store", "restore"},
+	{"store", "prune"},
+	{"store", "export"},
 }
 
 // IsAppRequired determines if a command requires a full application to be built by