+++ store/v2/CHANGELOG.md
@@ -26,9 +26,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## [Unreleased]
 
 ### Improvements
-
+* [#23568](https://github.com/cosmos/cosmos-sdk/pull/23568) Remove auto migration and fix restore cmd
 * [#23013](https://github.com/cosmos/cosmos-sdk/pull/23013) Support memDB for sims
 
+### Bug Fixes
+
+* [#23552](https://github.com/cosmos/cosmos-sdk/pull/23552) Fix pebbleDB integration
 
 ## [v2.0.0-beta.2](https://github.com/cosmos/cosmos-sdk/releases/tag/store/v2.0.0-beta.2)